lesson_25/app-postgress/requirements.txt

flask
psycopg2-binary
